Rebuild of Tesseract and Leptonica for OS X for Lion and Mountain Lion compatibility
I've rebuilt the Tesseract and Leptonica libraries with a '-mmacosx-version-min=10.7' switch that should make them compatible with OS X 10.7 (Lion), OS X 10.8 (Mountain Lion) an OS X 10.9 (Mavericks). I expect that this will fix crashes Lion and Mountain Lion such as #189

I also took the opportunity to move the Tesseract data to within the OS X application bundle so (for OS X) it doesn't need to copied to a temp directory on startup.