Updater no longer dumps instructions in /Applications folder
A file called instructions-osx.txt was being saved in the /Applications folder (or wherever you store your HearthStats.app) that shouldn't be. The file was intended as instructions for people downloading the first time, but it was also being extracted during updates and placed in that folder. I've changed the updater to exclude that file when updating an existing copy of HearthStats.