Fixed location of log file on OS X
The OS X build was accidentally including the username of the person who builds the HearthStats.net Uploader in the path of the log file. That meant it wouldn't log anything unless you built it yourself! So I've removed the username from the build, it will now log correctly for everyone.