Automated tests for OCR
Added some JUnit tests for OCR. These will automatically run OCR on the provided rank level and opponent name images to see if they are interpreted correctly. At the moment some of the tests are failing, so that means there's some work to do on the OCR to make it more accurate! :flushed: