Improved detection of victory and defeat screen
Implemented a new technique to detect the victory and defeat screens that better copes with the movement of the popup. This means it detects the result much faster -- pretty much as soon as the popup appears. That should make it very hard to dismiss the victory or defeat screen before the HearthStats.net Uploader has had a chance to detect the result.

Also added another JUnit test, this time to test how effective the victory and defeat detection is.