Much more accurate rank detection
Changed the way the screenshot of the rank is filtered so that only the number is extracted and all background is completely removed. This results in an almost perfect image for OCR, so the rank detection is extremely accurate now.

As a bonus the new filtering method is way faster than the old one, at least ten times faster!