Fixed detection of ranks 1 to 9
Ranks 1 to 9 weren't being detected because the OCR library Tesseract was ignoring any single-character words it found. That meant Tesseract treated ranks 1 to 9 as being blank. I've changed the configuration of Tesseract so that it will accept both single- and multiple-character words. Now all ranks from 1-25 are detected correctly (though legendary is probably not... I don't have a screenshot of that to work with!)