Save screenshot of hero/class when detection fails
Golden heroes currently aren't being detected (see #278) and we don't have the data we need to work out how to detect them because I don't have an golden heroes myself! This code saves several screenshots of a hero if its detection fails, so that the user can send the screenshot to use for analysis later.

As part of this change I also changed some existing pixel locations from a 1024x768 to a 1600x1200 basis so that all locations are expressed in consistent terms.