Increased maximum memory JVM setting
The HearthStats Uploader is currently running out of memory on Windows if it is left running for a few hours (see #264 and #279). This may be a memory leak, but until that can be determined I have increased the maximum memory allowed for both Windows and OS X to reduce the risk of a failure due to running out of memory.