Added match result popup
Implement a match result popup at the end of each match, as suggested in #252.

The new match result popup will appear automatically if there is any information missing from a match, so that you can fix up any detection errors yourself. There is also a new setting on the Options tab that lets you always display the popup after a match, or never show it if you prefer.