New implementation of the automatic updater
This new updater makes requests directly into the GitHub API to identify the latest version, instead of requiring us to manually update the version files and check them it. It's also implemented as a separate module, allowing it to be run while the HearthStats Uploader is closed and thus avoid problems where the uploader isn't updated properly because it's in use.

Note that I haven't yet implemented most of the necessary error handling, so it will currently throw exceptions if there are no new releases published (which is the case right now...)