Changed name from HearthStats Uploader to HearthStats Companion
This is only a UI change so far. Changing the name of the project on GitHub and the name of the output .jar files would break the updater, so we need to make changes to the updater to support both names before we can complete the move from to the new name.