Your hero class for a match is now taken from the deck by default
Previously HearthStats Companion would rely on screen detection to determine what hero you were playing. If there was a detection problem, such as a new golden hero being played, then the app wouldn't set your class for that match. That seemed silly because we already know the class from the deck you are playing.

So now your hero class for a match is set to whatever class the deck is. However it still performs the screen detection, so it's just a backup in case there's a problem with detection. The screen detection is still useful catching mistakes, such as when you've got a deck in the wrong slot.