Constructing a deck will no longer scroll the deck if you're previously exported a deck.
The deck export popup was not being removed properly when closed, which resulted in it attempt to detect a deck when you construct a deck. This resulted in it scrolling the empty deck just before it is constructed. The popup is now removed correctly so that it doesn't kick in during the deck construction.