better log file handling (#564) and more reliable coin/turn detection
actually this reverts partly the hack from 
9b01785581f20625f3c46e6a7751b82f3900b474

and replaces it with game start detection. It's not working reliably
yet but has no more the negative effects of messing up with
turn/coin detection (missing log events at the start of a game).

This commit also fixes the FileObserver :
* the file is not kept open, so HS can write in it even if started 
after the companion
* the FileObserver only tails the file if HS is detected