Fixed OutOfMemoryError on OS X on large retina screens
When playing full screen on a Retina iMac, the 5120x2880 screenshots were taking too much memory and causing HearthStats Companion to fail with an OutOfMemoryError. The Companion was storing up to 20 images in a buffer for video, which is unnecessary for OS X that doesn't yet have video recording, so I've stopped it buffering when not recording a video. This fixes the OutOfMemoryError.